requirements.txt
datasets>=2.14.0
transformers>=4.30.0
torch>=2.0.0
faiss-cpu>=1.7.4
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
scikit-learn>=1.3.0
huggingface-hub>=0.16.0

